implement marshal for config which has unmarshal, and add unit-test
