+++ pkg/api/v2/types.go
@@ -18,9 +18,14 @@
 package v2
 
 import (
+	"fmt"
+	"io/ioutil"
 	"net"
+	"os"
+	"path"
 	"time"
 
+	"github.com/alipay/sofa-mosn/pkg/utils"
 	"istio.io/api/mixer/v1/config/client"
 )
 
@@ -99,10 +104,41 @@ type HealthCheck struct {
 	IntervalJitter time.Duration `json:"-"`
 }
 
+// Marshal implement a json.Marshaler
+func (hc HealthCheck) MarshalJSON() (b []byte, err error) {
+	hc.HealthCheckConfig.IntervalConfig.Duration = hc.Interval
+	hc.HealthCheckConfig.IntervalJitterConfig.Duration = hc.IntervalJitter
+	hc.HealthCheckConfig.TimeoutConfig.Duration = hc.Timeout
+	return json.Marshal(hc.HealthCheckConfig)
+}
+
+func (hc *HealthCheck) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &hc.HealthCheckConfig); err != nil {
+		return err
+	}
+	hc.Timeout = hc.TimeoutConfig.Duration
+	hc.Interval = hc.IntervalConfig.Duration
+	hc.IntervalJitter = hc.IntervalJitterConfig.Duration
+	return nil
+}
+
 // Host represenets a host information
 type Host struct {
 	HostConfig
 	MetaData Metadata `json:"-"`
+}
+
+func (h Host) MarshalJSON() (b []byte, err error) {
+	h.HostConfig.MetaDataConfig = metadataToConfig(h.MetaData)
+	return json.Marshal(h.HostConfig)
+}
+
+func (h *Host) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &h.HostConfig); err != nil {
+		return err
+	}
+	h.MetaData = configToMetadata(h.MetaDataConfig)
+	return nil
 }
 
 // Listener contains the listener's information
@@ -139,10 +175,36 @@ type HealthCheckFilter struct {
 	CacheTime time.Duration `json:"-"`
 }
 
+func (hf HealthCheckFilter) MarshalJSON() (b []byte, err error) {
+	hf.HealthCheckFilterConfig.CacheTimeConfig.Duration = hf.CacheTime
+	return json.Marshal(hf.HealthCheckFilterConfig)
+}
+
+func (hf *HealthCheckFilter) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &hf.HealthCheckFilterConfig); err != nil {
+		return err
+	}
+	hf.CacheTime = hf.CacheTimeConfig.Duration
+	return nil
+}
+
 // FaultInject
 type FaultInject struct {
 	FaultInjectConfig
 	DelayDuration uint64 `json:"-"`
+}
+
+func (f FaultInject) Marshal() (b []byte, err error) {
+	f.FaultInjectConfig.DelayDurationConfig.Duration = time.Duration(f.DelayDuration)
+	return json.Marshal(f.FaultInjectConfig)
+}
+
+func (f *FaultInject) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &f.FaultInjectConfig); err != nil {
+		return err
+	}
+	f.DelayDuration = uint64(f.DelayDurationConfig.Duration)
+	return nil
 }
 
 // StreamFaultInject
@@ -157,6 +219,20 @@ type DelayInject struct {
 	DelayInjectConfig
 	Delay time.Duration `json:"-"`
 }
+
+func (d DelayInject) Marshal() (b []byte, err error) {
+	d.DelayInjectConfig.DelayDurationConfig.Duration = d.Delay
+	return json.Marshal(d.DelayInjectConfig)
+}
+
+func (d *DelayInject) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &d.DelayInjectConfig); err != nil {
+		return err
+	}
+	d.Delay = d.DelayDurationConfig.Duration
+	return nil
+}
+
 type AbortInject struct {
 	Status  int    `json:"status,omitempty"`
 	Percent uint32 `json:"percentage,omitempty"`
@@ -174,11 +250,39 @@ type Router struct {
 	Metadata Metadata `json:"-"`
 }
 
+func (r Router) MarshalJSON() (b []byte, err error) {
+	r.RouterConfig.MetadataConfig = metadataToConfig(r.Metadata)
+	return json.Marshal(r.RouterConfig)
+}
+
+func (r *Router) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &r.RouterConfig); err != nil {
+		return err
+	}
+	r.Metadata = configToMetadata(r.MetadataConfig)
+	return nil
+}
+
 // RouteAction represents the information of route request to upstream clusters
 type RouteAction struct {
 	RouterActionConfig
 	MetadataMatch Metadata      `json:"-"`
 	Timeout       time.Duration `json:"-"`
+}
+
+func (r RouteAction) MarshalJSON() (b []byte, err error) {
+	r.RouterActionConfig.MetadataConfig = metadataToConfig(r.MetadataMatch)
+	r.RouterActionConfig.TimeoutConfig.Duration = r.Timeout
+	return json.Marshal(r.RouterActionConfig)
+}
+
+func (r *RouteAction) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &r.RouterActionConfig); err != nil {
+		return err
+	}
+	r.Timeout = r.RouterActionConfig.TimeoutConfig.Duration
+	r.MetadataMatch = configToMetadata(r.MetadataConfig)
+	return nil
 }
 
 // Decorator
@@ -192,16 +296,50 @@ type ClusterWeight struct {
 	MetadataMatch Metadata `json:"-"`
 }
 
+func (cw ClusterWeight) MarshalJSON() (b []byte, err error) {
+	cw.ClusterWeightConfig.MetadataConfig = metadataToConfig(cw.MetadataMatch)
+	return json.Marshal(cw.ClusterWeightConfig)
+}
+
+func (cw *ClusterWeight) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &cw.ClusterWeightConfig); err != nil {
+		return err
+	}
+	cw.MetadataMatch = configToMetadata(cw.MetadataConfig)
+	return nil
+}
+
 // RetryPolicy represents the retry parameters
 type RetryPolicy struct {
 	RetryPolicyConfig
 	RetryTimeout time.Duration `json:"-"`
 }
 
+func (rp RetryPolicy) MarshalJSON() (b []byte, err error) {
+	rp.RetryPolicyConfig.RetryTimeoutConfig.Duration = rp.RetryTimeout
+	return json.Marshal(rp.RetryPolicyConfig)
+}
+
+func (rp *RetryPolicy) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &rp.RetryPolicyConfig); err != nil {
+		return err
+	}
+	rp.RetryTimeout = rp.RetryTimeoutConfig.Duration
+	return nil
+}
+
 // CircuitBreakers is a configuration of circuit breakers
 // CircuitBreakers implements json.Marshaler and json.Unmarshaler
 type CircuitBreakers struct {
 	Thresholds []Thresholds
+}
+
+// CircuitBreakers's implements json.Marshaler and json.Unmarshaler
+func (cb CircuitBreakers) MarshalJSON() (b []byte, err error) {
+	return json.Marshal(cb.Thresholds)
+}
+func (cb *CircuitBreakers) UnmarshalJSON(b []byte) (err error) {
+	return json.Unmarshal(b, &cb.Thresholds)
 }
 
 type Thresholds struct {
@@ -263,6 +401,34 @@ type FilterChain struct {
 	TLSContexts []TLSConfig `json:"-"`
 }
 
+func (fc FilterChain) MarshalJSON() (b []byte, err error) {
+	if len(fc.TLSContexts) > 0 { // use tls_context_set
+		fc.TLSConfig = nil
+		fc.TLSConfigs = fc.TLSContexts
+	}
+	return json.Marshal(fc.FilterChainConfig)
+}
+
+func (fc *FilterChain) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &fc.FilterChainConfig); err != nil {
+		return err
+	}
+	if fc.TLSConfig != nil && len(fc.TLSConfigs) > 0 {
+		return ErrDuplicateTLSConfig
+	}
+	if len(fc.TLSConfigs) > 0 {
+		fc.TLSContexts = make([]TLSConfig, len(fc.TLSConfigs))
+		copy(fc.TLSContexts, fc.TLSConfigs)
+	} else { // no tls_context_set, use tls_context
+		if fc.TLSConfig == nil { // no tls_context, generate a default one
+			fc.TLSContexts = append(fc.TLSContexts, TLSConfig{})
+		} else { // use tls_context
+			fc.TLSContexts = append(fc.TLSContexts, *fc.TLSConfig)
+		}
+	}
+	return nil
+}
+
 // Filter is a config to make up a filter
 type Filter struct {
 	Type   string                 `json:"type,omitempty"`
@@ -314,6 +480,81 @@ type HeaderValue struct {
 type RouterConfiguration struct {
 	VirtualHosts []*VirtualHost `json:"-"`
 	RouterConfigurationConfig
+}
+
+// Marshal memory config into json, if dynamic mode is configured, write json file
+func (rc RouterConfiguration) MarshalJSON() (b []byte, err error) {
+	if rc.RouterConfigPath == "" {
+		rc.StaticVirtualHosts = rc.VirtualHosts
+		return json.Marshal(rc.RouterConfigurationConfig)
+	}
+	// dynamic mode, should write file
+	// first, get all the files in the directory
+	files, err := ioutil.ReadDir(rc.RouterConfigPath)
+	if err != nil {
+		return nil, err
+	}
+	allFiles := make(map[string]struct{}, len(files))
+	for _, f := range files {
+		allFiles[f.Name()] = struct{}{}
+	}
+	// file name is virtualhost name, if not exists, use {unixnano}.json
+	for _, vh := range rc.VirtualHosts {
+		fileName := vh.Name
+		if fileName == "" {
+			fileName = fmt.Sprintf("%d", time.Now().UnixNano())
+		}
+		data, err := json.MarshalIndent(vh, "", " ")
+		if err != nil {
+			return nil, err
+		}
+		fileName = fileName + ".json"
+		delete(allFiles, fileName)
+		fileName = path.Join(rc.RouterConfigPath, fileName)
+		if err := utils.WriteFileSafety(fileName, data, 0644); err != nil {
+			return nil, err
+		}
+	}
+	// delete invalid files
+	for f := range allFiles {
+		os.Remove(path.Join(rc.RouterConfigPath, f))
+	}
+	return json.Marshal(rc.RouterConfigurationConfig)
+}
+
+func (rc *RouterConfiguration) UnmarshalJSON(b []byte) error {
+	if err := json.Unmarshal(b, &rc.RouterConfigurationConfig); err != nil {
+		return err
+	}
+	cfg := rc.RouterConfigurationConfig
+	// only one of the config should be exists
+	if len(cfg.StaticVirtualHosts) > 0 && cfg.RouterConfigPath != "" {
+		return ErrDuplicateStaticAndDynamic
+	}
+	if len(cfg.StaticVirtualHosts) > 0 {
+		rc.VirtualHosts = cfg.StaticVirtualHosts
+	}
+	// Traversing path and parse the json
+	// assume all the files in the path are available json file, and no sub path
+	if cfg.RouterConfigPath != "" {
+		files, err := ioutil.ReadDir(cfg.RouterConfigPath)
+		if err != nil {
+			return err
+		}
+		for _, f := range files {
+			fileName := path.Join(cfg.RouterConfigPath, f.Name())
+			data, err := ioutil.ReadFile(fileName)
+			if err != nil {
+				return err
+			}
+			vh := &VirtualHost{}
+			if err := json.Unmarshal(data, vh); err != nil {
+				return err
+			}
+			rc.VirtualHosts = append(rc.VirtualHosts, vh)
+		}
+	}
+	return nil
 }
 
 // VirtualHost is used to make up the route table
@@ -382,6 +623,13 @@ type ApplicationInfo struct {
 // PublishInfo implements json.Marshaler and json.Unmarshaler
 type PublishInfo struct {
 	Pub PublishContent
+}
+
+func (pb PublishInfo) MarshalJSON() (b []byte, err error) {
+	return json.Marshal(pb.Pub)
+}
+func (pb *PublishInfo) UnmarshalJSON(b []byte) (err error) {
+	return json.Unmarshal(b, &pb.Pub)
 }
 
 type PublishContent struct {
